refactor: split per-package apply layer into _package.py
Move PackageContribution, build_contribution, shadowed_paths, _apply_payload,
_contains, and _under_any from _refresh.py into a new _package.py module.

_refresh.py now hosts the per-file pipeline only (enumerate, parse, observe,
cache); _package.py owns the per-package apply step that consumes payloads.
Pure file move with no behavior change.
